refactor: give the sled DataHeader key space its own value type
# Summary

`Header` no longer implements `SledSerde`. The legacy sled encoding moves
onto a `SledHeader` newtype in `sled_compat`, so the header type itself
carries no sled knowledge.

# Details

The header is live data — V004 keeps it in a plain `df_meta/VERSION`
file, read and written with `serde_json`. Only its presence in the
`DataHeader` sled key space is compatibility surface, and that is the
sole reason the type implemented a sled trait.

Separating them is also what makes the type movable: once `Header` lives
in `raft-config`, `store-compat` could not implement `SledSerde` for it
at all, since both trait and type would be foreign to it.

`SMEntry::DataHeader` and `RaftStoreEntry::DataHeader` pick up the new
value type on their own — both spell it as
`<DataHeader as SledKeySpace>::V`. A newtype renders as the bare inner
value in serde, so the frozen export format is unchanged; the exact JSON
is asserted by `metasrv_grpc_export::test_export`, which was confirmed to
fail on a deliberate change to that entry.
